Provide a built in public EmptyWireLoggerFactory class
